refactored megamoth conditional

+++ app/controllers/gatherings_controller.rb
@@ -11,12 +11,8 @@ class GatheringsController < ApplicationController
     leader = Leader.find_by(id: params[:leader_id])
     user = User.find_by(id: params[:user_id])
     gathering = Gathering.find_by(leader_id: leader.id, user_id: user.id)
-    if gathering
-      gathering.destroy
-      gathering = Gathering.create(leader_id: leader.id, user_id: user.id, query: params[:query], leader_handle: leader.handle) if (params[:query] != "")
-    else
-      gathering = Gathering.create(leader_id: leader.id, user_id: user.id, query: params[:query], leader_handle: leader.handle) unless params[:query] == ""
-    end
+    gathering.destroy if gathering
+    gathering = Gathering.create(leader_id: leader.id, user_id: user.id, query: params[:query], leader_handle: leader.handle) unless params[:query] == ""
 
     render json: {gathering: gathering}
   end
